requirements.txt
Flask
Flask-Dance
requests
beautifulsoup4
openai
python-dotenv
